fix(stock-prep): keep policy-honesty tests off a frozen input
`plugins/plugin-integration-core/package.json` is a PINNED file in the sealed-export
package-provenance frozen manifest (PINNED_RUNTIME_FILES). Appending the new suite to
`scripts.test` changed its digest and REDed `sealed-export-package-provenance.test.cjs`,
taking `integration-guard` with it.

Proven by bisection, not inference: restoring only package.json to its base content makes
`sealed-export-package-provenance.test.cjs` pass again. The same run proves the three
lib/ modules this PR edits are NOT pinned.

Fix: revert package.json to byte-identical-with-main and relocate the policy-honesty
tests into `stock-preparation-conflict-policies.test.cjs`, which is already in the
`scripts.test` chain. Same CI coverage, zero frozen inputs touched. The two-point wiring
guard is dropped with the wiring it guarded.
